refactor: divide review
