+++ src/App.tsx
@@ -3,11 +3,11 @@ import { Provider } from "react-redux";
 import { BrowserRouter, Route, Routes } from "react-router-dom";
 import { PersistGate } from "redux-persist/integration/react";
 import persistStore from "redux-persist/es/persistStore";
-import HospitalList from "./components/pages/HospitalList";
+import HospitalPage from "./components/pages/HospitalPage";
 import ReviewDetail from "./components/pages/ReviewDetail";
-import ReviewList from "./components/pages/ReviewList";
 import GlobalStyle from "./styles/GlobalStyle";
 import store from "./store/store";
+import ReviewPage from "./components/pages/ReviewPage";
 
 const persistor = persistStore(store);
 const App = (): JSX.Element => {
@@ -17,10 +17,10 @@ const App = (): JSX.Element => {
         <GlobalStyle />
         <BrowserRouter>
           <Routes>
-            <Route path="/" element={<HospitalList />} />
-            <Route path="/review" element={<ReviewList />} />
+            <Route path="/" element={<HospitalPage />} />
+            <Route path="/review" element={<ReviewPage />} />
             <Route path="/detail" element={<ReviewDetail />} />
-            <Route path="*" element={<HospitalList />} />
+            <Route path="*" element={<HospitalPage />} />
           </Routes>
         </BrowserRouter>
       </PersistGate>
