+++ src/components/reviews/review/SuggestText.tsx
@@ -1,15 +1,6 @@
 import React from "react";
-import styled from "styled-components";
 import theme from "../../../styles/theme";
-
-type TextStyleProps = {
-  suggest: boolean;
-};
-
-const TextStyle = styled.span<TextStyleProps>`
-  color: ${(props): string =>
-    props.suggest ? theme.colors.purple : theme.colors.gray};
-`;
+import Span from "../../utils/Span";
 
 interface SuggestTextProps {
   suggest: boolean;
@@ -19,9 +10,12 @@ const SuggestText = (props: SuggestTextProps): JSX.Element => {
   const { suggest } = props;
 
   return (
-    <TextStyle suggest={suggest}>
+    <Span
+      fontSize="0.8rem"
+      color={suggest ? theme.colors.purple : theme.colors.gray}
+    >
       {suggest ? `재방문 의사 있음` : `재방문 의사 없음`}
-    </TextStyle>
+    </Span>
   );
 };
